+++ src/rai_core/rai/communication/ros2/connectors/__init__.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 from .action_mixin import ROS2ActionMixin
-from .ari_connector import ROS2Connector
+from .connector import ROS2Connector
 from .hri_connector import ROS2HRIConnector
 from .service_mixin import ROS2ServiceMixin
